support netstandard 1.6
