+++ UniqueIdGenerator.NetCore/UniqueIdGenerator.NetCore.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>net462;netstandard1.6</TargetFrameworks>
+    <AssemblyName>UniqueIdGenerator.NetCore</AssemblyName>
+    <PackageId>UniqueIdGenerator.NetCore</PackageId>
+    <PackageTags>netcore;snowflake;unique id;</PackageTags>
+    <Authors>UniqueIdGenerator.NetCore</Authors>
+    <Description>unique id generator</Description>
+    <PackageProjectUrl>https://github.com/mschuler/UniqueIdGenerator</PackageProjectUrl>
+    <Version>1.1.0</Version>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
+  </PropertyGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.6'">
+    <PackageReference Include="System.Threading.Thread" Version="4.3.0" />
+  </ItemGroup>
+</Project>
