+++ UniqueIdGenerator.net.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UniqueIdGenerator.Net", "IdGenerator.net\UniqueIdGenerator.Net.csproj", "{72288941-1E9C-4B91-8770-2408F6F27260}"
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.16
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UniqueIdGenerator.NetCore", "UniqueIdGenerator.NetCore\UniqueIdGenerator.NetCore.csproj", "{AD16CB05-D8E6-4E91-9ABC-66AB5D9BD7B4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UniqueIdGenerator.Net.Tests", "IdGenerator.net.Tests\UniqueIdGenerator.Net.Tests.csproj", "{ABC25329-4B08-4B81-9C5C-091C1F0AB7ED}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UniqueIdGenerator.NetCore.Tests", "UniqueIdGenerator.NetCore.Tests\UniqueIdGenerator.NetCore.Tests.csproj", "{1590C428-AF21-47DC-91D5-094D52ACED56}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{72288941-1E9C-4B91-8770-2408F6F27260}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{72288941-1E9C-4B91-8770-2408F6F27260}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{72288941-1E9C-4B91-8770-2408F6F27260}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{72288941-1E9C-4B91-8770-2408F6F27260}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ABC25329-4B08-4B81-9C5C-091C1F0AB7ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ABC25329-4B08-4B81-9C5C-091C1F0AB7ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ABC25329-4B08-4B81-9C5C-091C1F0AB7ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ABC25329-4B08-4B81-9C5C-091C1F0AB7ED}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AD16CB05-D8E6-4E91-9ABC-66AB5D9BD7B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD16CB05-D8E6-4E91-9ABC-66AB5D9BD7B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD16CB05-D8E6-4E91-9ABC-66AB5D9BD7B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD16CB05-D8E6-4E91-9ABC-66AB5D9BD7B4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1590C428-AF21-47DC-91D5-094D52ACED56}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1590C428-AF21-47DC-91D5-094D52ACED56}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1590C428-AF21-47DC-91D5-094D52ACED56}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1590C428-AF21-47DC-91D5-094D52ACED56}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
